sql: remove CalledOnNullInput field of tree.RoutineExpr
This commit removes the field `CalledOnNullInput` of `tree.RoutineExpr`.
Strict UDFs are now wrapped in a `CASE` expression so that they are not
invoked when any of the arguments are NULL. This allows for the removal
of special logic in evaluation of `tree.RoutineExpr`. This is one step
towards the goal of implementing UDFs with more generalized building
blocks. Ultimately, the special-purpose `memo.UDFExpr` should be
replaced with a general-purpose expression that can execute multiple
relational expression in sequence.

As a side effect of this change, strict UDFs will be folded to NULL at
optimization-time if they are presented with constant NULL arguments.
This is made possible by existing normalization rules.

Epic: None

Release note: None