sql: use optimizer-driven estimate for allocation size
This commit plumbs the estimated row count for the number of rows
produced by the GroupBy from the optimizer all the way to the execution,
and then it uses this estimate to come up with the initial allocation
size in the hash aggregator in the vectorized engine (mentioned in the
previous commit).

As expected, benchmarks (assuming the perfect optimizer estimation) show
noticeable improvement in some cases, both in speed and size of
allocations, slightly increasing the number of allocated objects in
a few cases: https://gist.github.com/yuzefovich/1546c9d133c6ba52cdda6ecd9297ba7d.

Release note: None