schemaexpr: refactor NameResolutionVisitor
The `NameResolutionVisitor` is not used outside the `schemaexpr`
package, so it has been unexported. Additionally, it no longer attempts
to rebind indexed vars or normalized function names. Neither of these
steps are necessary because this struct is only used to resolved column
names to indexed vars for computed column and partial index expressions.

Release note: None