opt: introduce memo.ExistsPrivate and add lazy projection column
This commit introduces the `memo.ExistsPrivate` struct, which wraps the
`memo.SubqueryPrivate` struct. This new private struct has a column that
is used during the execbuilder phase if the Exists subquery is planned
as a lazily-evaluated routine. By adding this column to the Exists
expression, we can avoid the anti-pattern of creating a new column ID in
the execbuilder. This will be beneficial in future commits in which
execbuilder will rely on knowing the maximum column ID for a query plan
before it begins constructing execution nodes.

Release note: None