sql: force PCR reader catalogs to use a fixed timestamp
Previously, the original logic for advancing the PCR reader catalog
timestamp ran into the risk of mixing different timestamps, since the
lease manager relies on range feeds. So, it was possible for a single
txn to observe some newer and older descriptors. To address this, this
patch does the following:

1) It introduces an assertion in the leased descriptors to ensure that
   all external row data tables have the same timestamp.
2) The connection executor is modified for PCR reader catalog to always
   use AOST queries, where the AOST timestamp is determine based on the
   range feeds check point. This ensures that all data up to the
   timestamp synced up. We detect PCR reader catalog by checking the
   the ReplicatedPCRVersion on the systme database, which will be set
   during bootstrap.

Fixes: #130812

Release note: None