opt: track stats timestamps in execbuilder instead of durations
Execbuilder now tracks the timestamps of stats collections and forecasts
instead of the duration between the current time and those timestamps.
This decouples execbuild-time from execution-time and will allow for
accurate instrumentation of time-since-stats-collection and
time-since-stats-forecast when `planNode`s are cached in the future—in
which case execbuild will only happen once for the query, not for every
execution.

Release note: None